Crate rustlings

Modules

exercise 🔒

run 🔒

ui 🔒

verify 🔒

Structs

Args 🔒

Rustlings is a collection of small exercises to get you used to writing and reading Rust code

HintArgs 🔒

Returns a hint for the given exercise

ListArgs 🔒

Lists the exercises available in Rustlings

RunArgs 🔒

Runs/Tests a single exercise

VerifyArgs 🔒

Verifies all exercises according to the recommended order

WatchArgs 🔒

Reruns verify when files were edited

Enums

Subcommands 🔒

WatchStatus 🔒

Constants

VERSION 🔒

Functions

find_exercise 🔒

main 🔒

rustc_exists 🔒

spawn_watch_shell 🔒

watch 🔒