Crate rustlings

Modules

exercise ğŸ”’

run ğŸ”’

ui ğŸ”’

verify ğŸ”’

Structs

Args ğŸ”’

Rustlings is a collection of small exercises to get you used to writing and reading Rust code

HintArgs ğŸ”’

Returns a hint for the given exercise

ListArgs ğŸ”’

Lists the exercises available in Rustlings

RunArgs ğŸ”’

Runs/Tests a single exercise

VerifyArgs ğŸ”’

Verifies all exercises according to the recommended order

WatchArgs ğŸ”’

Reruns verify when files were edited

Enums

Subcommands ğŸ”’

WatchStatus ğŸ”’

Constants

VERSION ğŸ”’

Functions

find_exercise ğŸ”’

main ğŸ”’

rustc_exists ğŸ”’

spawn_watch_shell ğŸ”’

watch ğŸ”’